Custom JSON adapter - mapping wrapper content to a list

+++ loldd/src/main/java/com/rabbit/orange/loldd/data/model/Stats.java
@@ -1,14 +1,14 @@
 package com.rabbit.orange.loldd.data.model;
 
 /**
- * Created by Łukasz on 10.02.2018.
+ * Champion stats
  */
 
 public class Stats {
     public float hp;
-    public int hpperlevel;
+    public float hpperlevel;
     public float mp;
-    public int mpperlevel;
+    public float mpperlevel;
     public int movespeed;
     public float armor;
     public float armorperlevel;
@@ -17,12 +17,12 @@ public class Stats {
     public int attackrange;
     public float hpregen;
     public float hpregenperlevel;
-    public int mpregen;
+    public float mpregen;
     public float mpregenperlevel;
     public int crit;
     public int critperlevel;
     public float attackdamage;
-    public int attackdamageperlevel;
+    public float attackdamageperlevel;
     public float attackspeedoffset;
     public float attackspeedperlevel;
 }
